Refactor MyServer
This is the purer approach where we don't actually code to a dynamic
interface. It uses static polymorphism instead to deal with the
message-sepcific handlers (OnMessage/OnMessageSent).

This kind of compile-time genericity (static polymorphism) simply does
not exist in C#. It can be tricky to get used to. There's also a
required trick to dispatch the member functions from the derived class:
CRTP ( Curiously Recurring Template Patern, which in the code can be
found as the `Derived` template argument together with the `derived()`
accessor function).

I'd suggest that in this case your interface is actually too tightly
coupled (with the callbacks) anyways to make it worthwhile, but it's
good to realize that C++ templates are ***fundamentally*** more powerful
than C# generics, don't let the syntax fool you.